stolen macbook air on 2018-02-26
The stolen laptop was breached into and alerts from my gmail indicated attempts at id theft. The waterloo regional police service naturally couldn't do anything about the issue. They operate under the assumption that if the "crime" did not involve damages worth 25k CAD in liquid assets the issue is too small to be bothered with. The idea of cyber crimes that can be committed by a perp having access to the victim's personal data doesn't register in the minds of an outdated policing service.